Split component for FullPathToSelf
