+++ src/File/FullPathToSelf.php
@@ -1,0 +1,50 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Php\Pie\File;
+
+use RuntimeException;
+
+use function array_key_exists;
+use function getcwd;
+use function is_string;
+use function preg_match;
+use function realpath;
+
+use const DIRECTORY_SEPARATOR;
+
+/** @internal This is not public API for PIE, so should not be depended upon unless you accept the risk of BC breaks */
+class FullPathToSelf
+{
+    /** @return non-empty-string */
+    public function __invoke(): string
+    {
+        /** @psalm-suppress TypeDoesNotContainType */
+        $phpSelf = array_key_exists('PHP_SELF', $_SERVER) && is_string($_SERVER['PHP_SELF']) ? $_SERVER['PHP_SELF'] : '';
+        if ($phpSelf === '') {
+            throw new RuntimeException('Could not find PHP_SELF');
+        }
+
+        return $this->isAbsolutePath($phpSelf)
+            ? $phpSelf
+            : (getcwd() . DIRECTORY_SEPARATOR . $phpSelf);
+    }
+
+    private function isAbsolutePath(string $path): bool
+    {
+        if (realpath($path) === $path) {
+            return true;
+        }
+
+        if ($path === '' || $path === '.') {
+            return false;
+        }
+
+        if (preg_match('#^[a-zA-Z]:\\\\#', $path)) {
+            return true;
+        }
+
+        return $path[0] === '/' || $path[0] === '\\';
+    }
+}
